Ship-readiness: updates, CI, and a log that can't grow or gossip
Three things stood between this and a release someone else could install.

Updates. Glyphio now checks GitHub on launch and can install a newer version
from Settings → About. Verification uses the project's own minisign key rather
than Apple's signature, which is the point: updates work identically on an
unsigned build, so fixes can ship before there's $99 for a Developer ID. A
Homebrew-managed copy is detected and handed `brew upgrade` instead — self-
updating underneath brew leaves it convinced it has a version it doesn't.

The check is the only network call an otherwise-unconfigured Glyphio makes, so
it is a setting, not an assumption, and README/SECURITY.md no longer claim we
make none.

CI. Clippy and tests on every push, for both workspaces. It cannot sign or
release — the certificate stays on one machine — so it doesn't pretend to; what
it does catch is version drift between the three files that must agree, a
missing updater key, and a private key that wandered into the repo. Clippy is
gated at -D warnings, which meant fixing the 13 lints already present and
migrating three deprecated shell().open calls to tauri-plugin-opener. No
cargo fmt gate: this codebase is hand-formatted wider than rustfmt's default
and reformatting it would rewrite every blame line to prove nothing.

Logging. The file is long-lived, lands in bug reports, and was 0644 and
unbounded. Now: 4 MB with one rollover, 0600 (fixed up in place for files an
older build created), per-line truncation so a runaway subprocess can't dump a
megabyte, and redaction of URLs down to scheme+host on the engine's piped
stderr — the only text in the log we don't write ourselves. The module header
states what must never be logged and why: page identity, snippet bodies, sync
tokens, clipboard contents.

Distribution without Apple. release.sh now detects a Developer ID and, if it
finds one, signs with it and notarizes; if not, it falls back to the self-signed
identity and says so plainly. Nothing needs editing the day that certificate
exists. Until then docs/INSTALL.md covers the three honest routes, with a
Homebrew cask (packaging/homebrew) as the one that avoids Gatekeeper entirely.

Editor windows now carry MoveToActiveSpace, so an editor left on another
desktop comes to you instead of dragging you to it. The full-screen case is
deliberately left alone: fixing it needs FullScreenAuxiliary, which is mutually
exclusive with the bit that gives a window its own full-screen button. Losing
that to win this is a bad trade, and a silent capture sidesteps it anyway.

docs/WINDOWS.md surveys the port: the espanso fork keeps upstream's win32
support so expansion is nearly free, ~1,200 lines of capture are not, and the
interactive picker — which macOS hands us for free — has no Windows equivalent
and is the long pole.

+++ src-tauri/examples/rect_probe.rs
@@ -61,8 +61,8 @@ fn main() {
             .with_excluding_windows(&[])
             .build();
         let config = SCStreamConfiguration::new()
-            .with_width(d.width() as u32)
-            .with_height(d.height() as u32);
+            .with_width(d.width())
+            .with_height(d.height());
         match SCScreenshotManager::capture_image(&filter, &config) {
             Ok(img) => {
                 let (w, h) = (img.width() as u32, img.height() as u32);
